chore : api -> controller로 test 디렉터리 위치 변경
