+++ src/test/java/com/kakao/golajuma/comment/web/controller/UpdateCommentApiTest.java
@@ -1,10 +1,12 @@
-package com.kakao.golajuma.comment.api;
+package com.kakao.golajuma.comment.web.controller;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.kakao.golajuma.auth.domain.token.TokenProvider;
 import com.kakao.golajuma.comment.persistence.entity.CommentEntity;
 import com.kakao.golajuma.comment.persistence.repository.CommentRepository;
+import com.kakao.golajuma.comment.web.dto.request.UpdateCommentRequest;
 import javax.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -21,10 +23,14 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
 @AutoConfigureMockMvc
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK)
-class DeleteCommentApiTest {
+class UpdateCommentApiTest {
+	@Autowired private ObjectMapper om;
+
 	@Autowired private MockMvc mvc;
+
 	@Autowired private TokenProvider tokenProvider;
 	@Autowired private CommentRepository commentRepository;
+
 	private String jwtToken;
 
 	@BeforeEach
@@ -33,62 +39,99 @@ class DeleteCommentApiTest {
 		jwtToken = tokenProvider.createAccessToken(1L);
 	}
 
-	@DisplayName("유저는 댓글을 삭제하는데 성공한다.")
+	@DisplayName("유저는 댓글을 업데이트 하는데 성공한다.")
 	@Transactional
 	@Test
-	void delete_comment_success_tes() throws Exception {
+	void updateTest() throws Exception {
 		// given
 		CommentEntity commentEntity = commentRepository.findByUserId(1L).stream().findFirst().get();
 		Long voteId = commentEntity.getVoteId();
 		Long commentId = commentEntity.getId();
+		UpdateCommentRequest requestDto = UpdateCommentRequest.builder().content("메롱이다.22").build();
+		String requestBody = om.writeValueAsString(requestDto);
+
 		// when
 		ResultActions resultActions =
 				mvc.perform(
-						MockMvcRequestBuilders.delete("/votes/" + voteId + "/comments/" + commentId)
+						MockMvcRequestBuilders.put("/votes/" + voteId + "/comments/" + commentId)
 								.header("Authorization", "Bearer " + jwtToken)
+								.content(requestBody)
 								.contentType(MediaType.APPLICATION_JSON));
+
 		// then
 		resultActions
 				.andExpect(MockMvcResultMatchers.status().isOk())
 				.andExpect(MockMvcResultMatchers.content().contentType("application/json"))
+				.andExpect(MockMvcResultMatchers.jsonPath("$.data.id").hasJsonPath())
+				.andExpect(MockMvcResultMatchers.jsonPath("$.data.isOwner").hasJsonPath())
+				.andExpect(MockMvcResultMatchers.jsonPath("$.data.username").hasJsonPath())
+				.andExpect(MockMvcResultMatchers.jsonPath("$.data.content").hasJsonPath())
 				.andExpect(MockMvcResultMatchers.jsonPath("$.message").hasJsonPath());
 	}
 
-	@DisplayName("유저는 댓글을 삭제하는데 실패한다.")
+	@DisplayName("유저는 댓글을 업데이트 하는데 실패한다.")
 	@Nested
-	class delete_comment_fail_case {
-		@DisplayName("본인에게 존재하지 않은 댓글을 삭제하기 때문에 실패한다.")
-		@Transactional
+	class update_comment_fail_case {
+		@DisplayName("유저는 아무런 댓글을 작성하지 않고 업데이트해서 실패한다.")
 		@Test
-		void not_owner_delete_comment_fail_test() throws Exception {
+		void update_empty_comment_fail_test() throws Exception {
 			// given
-			CommentEntity commentEntity = commentRepository.findByUserId(2L).stream().findFirst().get();
+			CommentEntity commentEntity = commentRepository.findByUserId(1L).stream().findFirst().get();
 			Long voteId = commentEntity.getVoteId();
 			Long commentId = commentEntity.getId();
+			UpdateCommentRequest requestDto = UpdateCommentRequest.builder().content("").build();
+			String requestBody = om.writeValueAsString(requestDto);
 
 			// when
 			ResultActions resultActions =
 					mvc.perform(
-							MockMvcRequestBuilders.delete("/votes/" + voteId + "/comments/" + commentId)
+							MockMvcRequestBuilders.put("/votes/" + voteId + "/comments/" + commentId)
 									.header("Authorization", "Bearer " + jwtToken)
+									.content(requestBody)
+									.contentType(MediaType.APPLICATION_JSON));
+			// then
+			resultActions.andExpect(status().isBadRequest());
+		}
+
+		@DisplayName("다른 유저의 댓글을 수정하기 때문에 실패한다.")
+		@Transactional
+		@Test
+		void not_owner_update_comment_fail_test() throws Exception {
+			// given
+			CommentEntity commentEntity = commentRepository.findByUserId(2L).stream().findFirst().get();
+			Long voteId = commentEntity.getVoteId();
+			Long commentId = commentEntity.getId();
+			UpdateCommentRequest requestDto = UpdateCommentRequest.builder().content("fail").build();
+			String requestBody = om.writeValueAsString(requestDto);
+
+			// when
+			ResultActions resultActions =
+					mvc.perform(
+							MockMvcRequestBuilders.put("/votes/" + voteId + "/comments/" + commentId)
+									.header("Authorization", "Bearer " + jwtToken)
+									.content(requestBody)
 									.contentType(MediaType.APPLICATION_JSON));
 			// then
 			resultActions.andExpect(status().isNotFound());
 		}
 
-		@DisplayName("없는 댓글을 삭제하기 때문에 실패한다.")
+		@DisplayName("없는 댓글을 수정하기 때문에 실패한다.")
 		@Transactional
 		@Test
-		void not_exist_comment_delete_fail_test() throws Exception {
+		void not_exist_comment_update_fail_test() throws Exception {
 			// given
 			CommentEntity commentEntity = commentRepository.findByUserId(1L).stream().findFirst().get();
 			Long voteId = commentEntity.getVoteId();
 			Long commentId = 999999L;
+			UpdateCommentRequest requestDto = UpdateCommentRequest.builder().content("fail").build();
+			String requestBody = om.writeValueAsString(requestDto);
+
 			// when
 			ResultActions resultActions =
 					mvc.perform(
-							MockMvcRequestBuilders.delete("/votes/" + voteId + "/comments/" + commentId)
+							MockMvcRequestBuilders.put("/votes/" + voteId + "/comments/" + commentId)
 									.header("Authorization", "Bearer " + jwtToken)
+									.content(requestBody)
 									.contentType(MediaType.APPLICATION_JSON));
 			// then
 			resultActions.andExpect(status().isNotFound());
