+++ src/test/java/com/kakao/golajuma/vote/web/controller/CloseVoteControllerTest.java
@@ -1,4 +1,4 @@
-package com.kakao.golajuma.vote.api;
+package com.kakao.golajuma.vote.web.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
