+++ src/test/java/com/kakao/golajuma/vote/web/controller/GetVotesControllerTest.java
@@ -1,4 +1,4 @@
-package com.kakao.golajuma.vote.api;
+package com.kakao.golajuma.vote.web.controller;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
